Revert "Merge pull request #14538 from cdapio/cherrypick/CDAP-19529-3-67"
This reverts commit 2dd4793a3747da23994006cf8e372238321cce69, reversing
changes made to 17da98a909e87ed87eec085eb55c35e5055eb309.
