+++ cdap-app-fabric/src/main/java/io/cdap/cdap/master/environment/DefaultMasterEnvironmentRunnableContext.java
@@ -16,24 +16,15 @@
 
 package io.cdap.cdap.master.environment;
 
-import io.cdap.cdap.app.runtime.ProgramRuntimeProvider;
-import io.cdap.cdap.common.conf.CConfiguration;
 import io.cdap.cdap.common.conf.Constants;
 import io.cdap.cdap.common.http.DefaultHttpRequestConfig;
 import io.cdap.cdap.common.internal.remote.RemoteClient;
 import io.cdap.cdap.common.internal.remote.RemoteClientFactory;
-import io.cdap.cdap.common.lang.CombineClassLoader;
-import io.cdap.cdap.internal.app.runtime.ProgramRuntimeProviderLoader;
 import io.cdap.cdap.master.spi.environment.MasterEnvironmentRunnableContext;
-import io.cdap.cdap.proto.ProgramType;
-import org.apache.twill.api.TwillRunnable;
 import org.apache.twill.filesystem.LocationFactory;
-import org.apache.twill.internal.utils.Instances;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 /**
  * Default implementation of {@link MasterEnvironmentRunnableContext}.
@@ -41,21 +32,13 @@ import java.util.stream.Collectors;
 public class DefaultMasterEnvironmentRunnableContext implements MasterEnvironmentRunnableContext {
   private final LocationFactory locationFactory;
   private final RemoteClient remoteClient;
-  private final CConfiguration cConf;
-  private final ProgramRuntimeProviderLoader programRuntimeProviderLoader;
-
-  private ClassLoader extensionCombinedClassLoader;
 
   public DefaultMasterEnvironmentRunnableContext(LocationFactory locationFactory,
-                                                 RemoteClientFactory remoteClientFactory,
-                                                 CConfiguration cConf) {
+                                                 RemoteClientFactory remoteClientFactory) {
     this.locationFactory = locationFactory;
     this.remoteClient = remoteClientFactory.createRemoteClient(
       Constants.Service.APP_FABRIC_HTTP,
       new DefaultHttpRequestConfig(false), "");
-    this.cConf = cConf;
-    this.programRuntimeProviderLoader = new ProgramRuntimeProviderLoader(cConf);
-    this.extensionCombinedClassLoader = null;
   }
 
   @Override
@@ -69,34 +52,5 @@ public class DefaultMasterEnvironmentRunnableContext implements MasterEnvironmen
   @Override
   public HttpURLConnection openHttpURLConnection(String resource) throws IOException {
     return remoteClient.openConnection(resource);
-  }
-
-  @Override
-  public TwillRunnable instantiateTwillRunnable(String className) {
-    Class<?> runnableClass;
-    try {
-      runnableClass = getClass().getClassLoader().loadClass(className);
-    } catch (ClassNotFoundException e) {
-      // Try loading the class from the runtime extensions.
-      if (extensionCombinedClassLoader == null) {
-        Map<ProgramType, ProgramRuntimeProvider> classLoaderProviderMap = programRuntimeProviderLoader.getAll();
-        extensionCombinedClassLoader = new CombineClassLoader(getClass().getClassLoader(),
-                                                              classLoaderProviderMap.entrySet().stream()
-                                                                .map(entry -> entry.getValue()
-                                                                  .createProgramClassLoader(cConf, entry.getKey()))
-                                                                .collect(Collectors.toList())
-                                                                .toArray(new ClassLoader[0]));
-      }
-      try {
-        runnableClass = extensionCombinedClassLoader.loadClass(className);
-      } catch (ClassNotFoundException cnfe) {
-        throw new RuntimeException(String.format("Failed to load twill runnable class from runtime extension '%s'",
-                                                 className), cnfe);
-      }
-    }
-    if (!TwillRunnable.class.isAssignableFrom(runnableClass)) {
-      throw new IllegalArgumentException("Class " + runnableClass + " is not an instance of " + TwillRunnable.class);
-    }
-    return (TwillRunnable) Instances.newInstance(runnableClass);
   }
 }
