+++ cdap-spark-core-base/src/test/java/io/cdap/cdap/app/runtime/spark/SparkProgramRuntimeProviderTest.java
@@ -16,46 +16,51 @@
 
 package io.cdap.cdap.app.runtime.spark;
 
+import com.google.inject.Injector;
 import io.cdap.cdap.app.runtime.ProgramRuntimeProvider;
 import io.cdap.cdap.common.conf.CConfiguration;
-import io.cdap.cdap.common.lang.ClassLoaders;
+import io.cdap.cdap.internal.app.worker.ConfiguratorTask;
+import io.cdap.cdap.internal.app.worker.SystemAppTask;
 import io.cdap.cdap.master.environment.MasterEnvironments;
 import io.cdap.cdap.master.spi.environment.MasterEnvironment;
 import io.cdap.cdap.proto.ProgramType;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import java.util.ServiceLoader;
 
 /**
- * Test for {@link SparkProgramRuntimeProvider}. We need to use the delegated test runner to ensure that there aren't
- * any casting errors for CDAP classes due to using different effective classloaders at runtime.
+ * Test for {@link SparkProgramRuntimeProvider}.
  */
-@RunWith(NoSparkClassLoaderTestRunner.class)
 public class SparkProgramRuntimeProviderTest {
 
   @Test
-  public void testSparkResourceFilterStreamingClassLoading() throws Exception {
-    // Tests the scenario in which the parent classloader of the SparkProgramRuntimeProvider does not have access to
-    // Spark classes. We simulate this by wrapping the entire test class with a delegated classloader which filters out
-    // Spark classes.
-
-    // Set the thread context classloader as a delegate classloader which does not have the Spark classes.
-    // This simulates that the parent classloader of SparkProgramRuntimeProvider does not have Spark classes.
-    ClassLoader noSparkClassLoader = getClass().getClassLoader();
-    ClassLoader tmp = ClassLoaders.setContextClassLoader(noSparkClassLoader);
+  public void testConfiguratorTaskInjector() throws Exception {
     MasterEnvironment masterEnvironment = new MockMasterEnvironment();
     masterEnvironment.initialize(null);
     MasterEnvironment tmpMasterEnv = MasterEnvironments.setMasterEnvironment(masterEnvironment);
 
-    CConfiguration cConf = CConfiguration.create();
-    ServiceLoader<ProgramRuntimeProvider> serviceLoader = ServiceLoader.load(ProgramRuntimeProvider.class,
-                                                                             noSparkClassLoader);
+    Injector injector = ConfiguratorTask.createInjector(CConfiguration.create());
+    ServiceLoader<ProgramRuntimeProvider> serviceLoader = ServiceLoader.load(ProgramRuntimeProvider.class);
     SparkProgramRuntimeProvider runtimeProvider = (SparkProgramRuntimeProvider) serviceLoader.iterator().next();
-    ClassLoader sparkClassLoader = runtimeProvider.createProgramClassLoader(cConf, ProgramType.SPARK);
-    // Load any spark streaming class.
-    sparkClassLoader.loadClass("org.apache.spark.streaming.StreamingContext");
+    runtimeProvider.createProgramRunner(ProgramType.SPARK, injector.getInstance(ProgramRuntimeProvider.Mode.class),
+                                        injector);
+
     MasterEnvironments.setMasterEnvironment(tmpMasterEnv);
-    ClassLoaders.setContextClassLoader(tmp);
   }
+
+  @Test
+  public void testSystemAppTaskInjector() throws Exception {
+    MasterEnvironment masterEnvironment = new MockMasterEnvironment();
+    masterEnvironment.initialize(null);
+    MasterEnvironment tmpMasterEnv = MasterEnvironments.setMasterEnvironment(masterEnvironment);
+
+    Injector injector = SystemAppTask.createInjector(CConfiguration.create());
+    ServiceLoader<ProgramRuntimeProvider> serviceLoader = ServiceLoader.load(ProgramRuntimeProvider.class);
+    SparkProgramRuntimeProvider runtimeProvider = (SparkProgramRuntimeProvider) serviceLoader.iterator().next();
+    runtimeProvider.createProgramRunner(ProgramType.SPARK, injector.getInstance(ProgramRuntimeProvider.Mode.class),
+                                        injector);
+
+    MasterEnvironments.setMasterEnvironment(tmpMasterEnv);
+  }
+
 }
